gen_requirements.txt
mypy-protobuf==2.4
# grpc-stubs==1.24.5
grpcio-tools==1.36.1